requirements.txt
h5py==2.6.0
matplotlib==2.0.0
numpy==1.19.5
scipy==0.18.1
tqdm==4.11.2
keras==2.0.2
scikit-learn==0.18.1
pillow==4.0.0
ipykernel==4.6.1
tensorflow==1.14.0
